requirements.txt
asgiref==3.6.0
black==23.3.0
certifi==2022.12.7
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
coverage==6.5.0
cryptography==40.0.2
defusedxml==0.7.1
distlib==0.3.6
Django==4.2
django-allauth==0.54.0
django-map-widgets==0.4.1
docopt==0.6.2
exceptiongroup==1.1.1
filelock==3.12.0
flake8==6.0.0
GDAL @ file:///D:/Projects/PlacesRemember/GDAL-3.4.3-cp310-cp310-win_amd64.whl
geographiclib==2.0
identify==2.5.23
idna==3.4
iniconfig==2.0.0
isort==5.12.0
mccabe==0.7.0
mypy-extensions==1.0.0
nodeenv==1.7.0
oauthlib==3.2.2
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.0
pluggy==1.0.0
pre-commit==3.2.2
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyJWT==2.6.0
pytest==7.3.1
python-dotenv==1.0.0
python3-openid==3.2.0
PyYAML==6.0
requests==2.29.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
tomli==2.0.1
tzdata==2023.3
urllib3==1.26.15
virtualenv==20.23.0
